utctf-2024/cryptordle/requirements.txt
pwntools
sagemath
